move Hero.js to home-page
