+++ src/components/home-page/Home.js
@@ -1,5 +1,5 @@
 import Header from "../header/Header";
-import Hero from "../hero/Hero";
+import Hero from "./hero/Hero";
 import WhyGruepr from "./why-gruepr/WhyGruepr.js";
 import Footer from "../footer/Footer.js";
 import TrustedSchools from "./trusted-schools/TrustedSchools";
